Add mesh deformation filter
The filter is based on CGAL::Surface_mesh_deformation.
The first input is the mesh to deform.
The second input is a vtkPointSet defining the target positions of
the points to move.
The third input (optional) is a vtkSelection defining a region of
interest containing points that can be moved during the deformation.